refactor: 🦄 moved folders to src
